Non converging update

+++ sigmoid.m
@@ -2,7 +2,7 @@ function [ neuralNet] = sigmoid(neuralNet)
 %UNTITLED9 Summary of this function goes here
 %   Detailed explanation goes here
 
-neuralNet = double(1.0) ./ ( double(1.0) + double(exp(-neuralNet)) );
+neuralNet = 1.0 ./ ( 1.0 + exp(-neuralNet) );
 
 end
